third_party/blink/web_tests/fast/forms/label/labels-set-htmlFor-attribute-expected.txt

Test that the labels list of a form control is dynamically updated when the htmlFor attribute of a label changes to point to that control.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check button
PASS element.labels.length is 0
PASS element.labels.length is 1
Check input
PASS element.labels.length is 0
PASS element.labels.length is 1
Check meter
PASS element.labels.length is 0
PASS element.labels.length is 1
Check output
PASS element.labels.length is 0
PASS element.labels.length is 1
Check progress
PASS element.labels.length is 0
PASS element.labels.length is 1
Check select
PASS element.labels.length is 0
PASS element.labels.length is 1
Check textarea
PASS element.labels.length is 0
PASS element.labels.length is 1
Check buttonType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check checkboxType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check colorType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check dateType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check datetimeType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check datetime-localType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check emailType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check fileType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check imageType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check monthType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check numberType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check passwordType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check radioType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check rangeType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check resetType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check searchType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check submitType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check telType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check textType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check timeType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check urlType
PASS element.labels.length is 0
PASS element.labels.length is 1
Check weekType
PASS element.labels.length is 0
PASS element.labels.length is 1
PASS successfullyParsed is true

TEST COMPLETE
